requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.2
colorama==0.4.3
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.0
Django==3.0.5
django-ckeditor==6.0.0
django-import-export==2.3.0
django-js-asset==1.2.2
et-xmlfile==1.0.1
filelock==3.0.12
get==2019.4.13
gunicorn==20.0.4
isort==4.3.21
jdcal==1.4.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.3
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.0.3
Pillow==7.2.0
post==2019.4.13
public==2019.4.13
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
query-string==2019.4.13
request==2019.4.13
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
virtualenv==20.0.18
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
